expand the sts class to include Poisson observations
